fix: address remaining clippy lints (cmp-owned, unnecessary-cast, manual-is-multiple-of)
- Fixed cmp_owned in ingest.rs by comparing with Path references
- Removed unnecessary-cast in uploader.rs
- Replaced manual modulo with is_multiple_of in render.rs

+++ src/services/uploader.rs
@@ -502,7 +502,7 @@ impl Uploader {
                     );
                 }
 
-                if n >= m_file_size as u64 {
+                if n >= m_file_size {
                     // Don't break, keep updating (likely 0 speeds) until aborted by main thread.
                     // This ensures we don't freeze on stale 'high speed' metrics if finalization takes time.
                 }
